9-binary_tree_height.c

+++ 9-binary_tree_height.c
@@ -11,8 +11,8 @@ size_t binary_tree_height(const binary_tree_t *tree)
 		return (0);
 	else
 	{
-		int left_height = binary_tree_height(tree->left);
-		int right_height = binary_tree_height(tree->right);
+		size_t left_height = binary_tree_height(tree->left);
+		size_t right_height = binary_tree_height(tree->right);
 
 		if (left_height >= right_height)
 			return (left_height + 1);
